🌸📇 Checkpoint ./src/models/collection.js:97325/490 ./src/utils/color.js:97325/540 ./src/components/collection/edit-collection-color-pop.js:97325/28 ./src/components/collection/collection-item-small.js:97325/28 ./src/components/collection/collection-item.js:97325/30 ./src/components/collection/note.js:97325/28 ./src/components/collection/container.js:97325/28
