+++ src/models/collection.js
@@ -1,6 +1,5 @@
 import { kebabCase } from 'lodash';
 import randomColor from 'randomcolor';
-import { hex as getHexContrastRatio } from 'wcag-contrast';
 
 import { CDN_URL } from 'Utils/constants';
 
@@ -11,17 +10,6 @@ import { getCollectionPair } from './words';
 
 export const FALLBACK_AVATAR_URL = 'https://cdn.glitch.com/1afc1ac4-170b-48af-b596-78fe15838ad3%2Fcollection-avatar.svg?1541449590339';
 export const defaultAvatar = 'https://cdn.glitch.com/1afc1ac4-170b-48af-b596-78fe15838ad3%2Fcollection-avatar.svg?1540389405633';
-
-export const getContrastWithLightText = (hex) => getHexContrastRatio(hex, '#fff');
-
-export const getContrastWithDarkText = (hex) => getHexContrastRatio(hex, '#222');
-
-export const isDarkColor = (hex) => {
-  if (!hex.startsWith('#')) return false;
-  const contrastWithLightText = getContrastWithLightText(hex);
-  const contrastWithDarkText = getContrastWithDarkText(hex);
-  return contrastWithLightText > contrastWithDarkText;
-};
 
 export function getAvatarUrl(id) {
   return `${CDN_URL}/collection-avatar/${id}.png`;
