+++ src/utils/color.js
@@ -1,5 +1,16 @@
 /* eslint-disable no-bitwise */
-/* eslint-disable import/prefer-default-export */
+import { hex as getHexContrastRatio } from 'wcag-contrast';
+
+export const getContrastWithLightText = (hex) => getHexContrastRatio(hex, '#fff');
+
+export const getContrastWithDarkText = (hex) => getHexContrastRatio(hex, '#222');
+
+export const isDarkColor = (hex) => {
+  if (!hex.startsWith('#')) return false;
+  const contrastWithLightText = getContrastWithLightText(hex);
+  const contrastWithDarkText = getContrastWithDarkText(hex);
+  return contrastWithLightText > contrastWithDarkText;
+};
 
 // from https://stackoverflow.com/a/21648508/1720985
 export const hexToRgbA = (hex) => {
